#Sql Added

+++ RocksAlexaRobot/modules/sql/global_bans_sql.py
@@ -3,18 +3,33 @@
 # Without Credit (Mother Fucker)
 # Rocks © @Dr_Asad_Ali © Rocks
 # Owner Asad + Harshit
-
+#  Copyright (C) 2017-2019, Paul Larsen
+#
+#  This program is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU Affero General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+#  GNU Affero General Public License for more details.
+#
+#  You should have received a copy of the GNU Affero General Public License
+#  along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 
 import threading
 
+from sqlalchemy import Column, UnicodeText, String, Boolean
+from sqlalchemy.sql.sqltypes import BigInteger
+
 from RocksAlexaRobot.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
 
 
 class GloballyBannedUsers(BASE):
     __tablename__ = "gbans"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     name = Column(UnicodeText, nullable=False)
     reason = Column(UnicodeText)
 
@@ -27,11 +42,7 @@ class GloballyBannedUsers(BASE):
         return "<GBanned User {} ({})>".format(self.name, self.user_id)
 
     def to_dict(self):
-        return {
-            "user_id": self.user_id,
-            "name": self.name,
-            "reason": self.reason
-        }
+        return {"user_id": self.user_id, "name": self.name, "reason": self.reason}
 
 
 class GbanSettings(BASE):
@@ -148,9 +159,7 @@ def num_gbanned_users():
 def __load_gbanned_userid_list():
     global GBANNED_LIST
     try:
-        GBANNED_LIST = {
-            x.user_id for x in SESSION.query(GloballyBannedUsers).all()
-        }
+        GBANNED_LIST = {x.user_id for x in SESSION.query(GloballyBannedUsers).all()}
     finally:
         SESSION.close()
 
@@ -159,9 +168,7 @@ def __load_gban_stat_list():
     global GBANSTAT_LIST
     try:
         GBANSTAT_LIST = {
-            x.chat_id
-            for x in SESSION.query(GbanSettings).all()
-            if not x.setting
+            x.chat_id for x in SESSION.query(GbanSettings).all() if not x.setting
         }
     finally:
         SESSION.close()
@@ -180,7 +187,3 @@ def migrate_chat(old_chat_id, new_chat_id):
 # Create in memory userid to avoid disk access
 __load_gbanned_userid_list()
 __load_gban_stat_list()
-
-
-# Roses are red, Violets are blue, A face like yours, Belongs in a zoo
-
